Add integration test for daily file in Data Directory
