+++ tests/unit/test_data_directory.py
@@ -7,7 +7,7 @@ from typing import Tuple
 from unittest import mock
 
 from requests.exceptions import HTTPError
-from tests import _mock_response
+from tests.unit import _mock_response
 
 from gmn_python_api import data_directory as gdd
 from gmn_python_api.meteor_summary_schema import _MODEL_TRAJECTORY_SUMMARY_FILE_PATH
